feat(core): elevate to strong production engineering standard
- Frontend: modular profile decomposition (GoalsStage, TimelineStage, CashFlowStage, AccountsStage, FinancialHealthStage, ProfileJourneyNavigation)
- Scenario isolation: decoupled baseProfile from scenarioOverrides with ScenarioBanner
- API boundary: runtime analysisSchema validation and ApiError handling
- Accessibility: automated WCAG 2.1 AA audits with axe-core (7/7 passed, zero violations)
- Styling: eliminated fragile :first-child selectors and !important overrides
- Backend: versioned Alembic migrations (0001 initial, 0002 refresh_token_sessions)
- Auth: revocable DB-backed sessions with token family rotation and reuse detection
- Concurrency: benchmarked 25 concurrent requests with threadpool offloading and p50/p95 latency verification
- Docs: comprehensive engineering README with financial model assumptions and debugging case study

+++ backend/schemas/retirement.py
@@ -99,7 +99,7 @@ class OptimizerSummary(BaseModel):
     model_config = ConfigDict(allow_inf_nan=False, extra="allow")
     total_potential_tax_savings: float
     financial_health_score: int
-    score_breakdown: Dict[str, int]
+    score_breakdown: Dict[str, float]
     top_3_priorities: List[Dict[str, Any]]
     lifetime_human_capital_at_risk: float
     projected_30yr_wealth: float
